requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.2
python-dotenv>=1.0.0
PyPDF2>=3.0.1
python-docx>=1.1.0
gTTS>=2.5.0
SpeechRecognition>=3.10.0
pydub>=0.25.1
pyaudio>=0.2.11